Fix (2nd attempt): Cannot download CSVs on IE11 (closes #121)
In the previous attempt I had neglected that we already used a patched version of `react-csv` that was modified to support large file sizes on Chrome. Depending on the browser I now import a differently patched version. We might still run into issues with large files on IE11, but I think it’s not worth to put much energy into this. As far as I understood we plan to switch to a more robust server side CSV generation in the future.